% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{dependent_search_terms}
\alias{dependent_search_terms}
\title{A helper function to construct dependent search terms for argument
\code{search_terms} of \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}}
\usage{
dependent_search_terms(...)
}
\arguments{
\item{...}{each element should be a formula, where the left hand side is a prerequisite
term(s) and the right hand side is one or more dependent terms that only appear
if the prerequisite(s) have already been selected.}
}
\description{
A helper function to construct dependent search terms for argument
\code{search_terms} of \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}
}
