% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{exclusive_search_terms}
\alias{exclusive_search_terms}
\title{A helper function to construct mutually exclusive sets of variables for
argument \code{search_terms} of \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}; if one of these terms
is selected, the remainder cannot be.}
\usage{
exclusive_search_terms(terms)
}
\arguments{
\item{terms}{a character vector of at least two search terms.}
}
\value{
A character vector that may be used as input for argument
\code{search_terms} of \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}.
}
\description{
A helper function to construct mutually exclusive sets of variables for
argument \code{search_terms} of \code{\link[=varsel]{varsel()}} or \code{\link[=cv_varsel]{cv_varsel()}}; if one of these terms
is selected, the remainder cannot be.
}
